Move TestCaseResult from test_results.py to results.py
